+++ src/EpicNewsletter_AU_AfternoonUpdate/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { canRender, COMPONENT_NAME } from './canRender';
 import { NewsletterEpic } from '../NewsletterEpic';
-import { NewsletterSubscribeCallback } from '../newsletterCommon/sharedComponents';
+import { NewsletterSubscribeCallback } from '../newsletterCommon';
 import type { TrackClick } from '../utils/tracking';
 
 // Image URL updated
